Finish merging files
