--- hq/home/apps.py
@@ -1,4 +1,0 @@
-from django.apps import AppConfig
-
-class HomeConfig(AppConfig):
-    name = 'home'
