+++ hq/homepage/apps.py
@@ -1,4 +1,4 @@
 from django.apps import AppConfig
 
 class HomeConfig(AppConfig):
-	name = 'homepage'
+    name = 'home'
